core:sandbox: Add support to block Module Auto-load Operations.
This is WIP.

Right now we only block explicit module load operations with
ProtectKernelModules=yes. This installs a seccomp on module syscalls and
removes CAP_SYS_MODULE from the capability bounding set of the unit. The
CAP_SYS_MODULE bit on its own is useless here since we already block the
related syscalls that need this capability with seccomp.

This protection only covers explicit module loading operations.
Automatic module loading due to a request for a feature that is not
loaded is not blocked. Privileged and unprivileged may trigger the kernel
to load extra modules, some modules include vulnerabilities like the
late CVE-2017-6074 DCCP double-free one.

There is a patchset under discussion that may allow to block this per
service:
http://www.openwall.com/lists/kernel-hardening/2017/04/09/1

Lets prepare for this now in case the feature is accepted.

With this change, ProtectKernelModules now support "no|yes|full|strict"

* "no" - Classic permissions, no protection

* "yes" - Use seccomp to block module syscalls and explicit module loading

* "full" - Use 'yes' protections + if the ModAutoRestrict LSM is
           available, then module auto-loading will be allowed only if
           the following capabilities CAP_SYS_MODULE or CAP_NET_ADMIN for
           modules with a 'netdev-%s' alias are set.

* "strict" - Use 'yes' protections + removes CAP_SYS_MODULE capability +
             if ModAutoRestrict LSM is available, then module
             auto-loading will be denied.

In case the Module Auto-load Restriction is not available in the kernel,
the only noticeable change here is now CAP_SYS_MODULE capability is only
removed on "strict" instead of "yes". This should have no effect since
we already did block explicit module loading by installing seccomp
filter on module syscalls. These syscalls are the only place where
CAP_SYS_MODULE was checked, so even if CAP_SYS_MODULE is there, the
syscalls are blocked. Also in this case 'full' will have the same protection
semantics as 'yes'.

With this change systemd-networkd will continue to work and device
modules that are loaded without checking any capability will continue to
work. Other services can be converted to have a "full" or "strict"
setting depending on the service.