path: add ReloadOnTrigger option. (#5770)
Changes triggered by .path units cause the target unit to be started.
However, this causes events to be lost: if the target service is running
and another file event triggers the .path unit, the next job start
request gets merged in the previously running one.

Add "ReloadOnTrigger" option to .path units. If enabled, the request to
the target unit is not "start", but "start_or_reload". In this case, if
the service is already started, future triggers to the .path unit will
cause reload events to the service. This limits the available service
types though, and can only be used with those services which can be
reloaded.