timesyncd: restore RTC when setting time from clock file
When timesyncd restores time from /var/lib/systemd/clock the system
presumably lost power which cleared the RTC. At this point, the RTC
is likely invalid. If there is no network connection, timesyncd
will not call clock_adjtime and the RTC will not get written with
a valid time again.

When the user shuts down the system and boot it later without
loosing power, the clock will be restored to /var/lib/systemd/clock
time again.

Writting the RTC at this point helps in this case: The RTC will be
able to run again, and on next start the time will have properly
advanced.

Signed-off-by: Stefan Agner <stefan@agner.ch>