execute: Return fork failures in exec_spawn()
In the event that an exec_spawn() fails to fork(), it will log the error,
but not return the failure to the calling function. This leads to an
assertion and freeze within service_spawn() as shown below:

    systemd: Failed to fork: Cannot allocate memory
    systemd: Assertion 'pid >= 1' failed at src/core/unit.c:1997, function unit_watch_pid(). Aborting.
    systemd: Caught <ABRT>, cannot fork for core dump: Cannot allocate memory
    systemd: Freezing execution.

By returning the failure, we avoid the assertion failure and subsequent
freezing event.