dhcp-server: stop routing the dhcp errors to sd_event loop
The DHCP server implementation will route all the errors related
to messages to sd_event loop. This will trigger sd_event loop to
disable the event handling, leaving the DHCP server in a mode where
requests are not handled anymore. In order to keep the DHCP server
functional, this patch avoids forwarding the errors to the sd_event
loop.

Signed-off-by: Nandor Han <nandor.han@ge.com>
Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.co.uk>