cryptsetup-generator: noauto instead of nofail should determine Before=
Currently the generator would add Before=cryptsetup.target unless nofail is set. This does not logical make sense since the option that determines whether the generated service would be pull by cryptsetup.target is noauto, while nofail merely determines whether it belongs to a .wants or a .requires.

generated services should have Before=cryptsetup.target even if it is merely wanted by cryptsetup.target (which is not currently the case).

Changing the nofail check to a noauto one. This should also prevent ordering cycle when a user wants a generated service to be pulled by a service that has to be started after cryptsetup.target, without having to set nofail in addition to noauto (which is not really logical and hence unexpected).