man: fix LD_LIBRARY_PATH example in environment.d
The example for LD_LIBRARY_PATH in the environment.d
man page is wrong.

When setting LD_LIBRARY_PATH, the new directory usually
needs to be at the front so it overrides old directories.

In the example, the colon delimiter is correctly prepended
to the front, but the actual new path is erroneously appended
to the end.

This commit moves it to the front where it belongs.