hwdb: Fix mapping for home-key on Cube iWork8 air
The intend of the hwdb entry for the Cube iWork8 air tablet buttons is
to fix them being swapped. The kernel is using left-meta for the
home key (it is being treated as the windows key on keyboards, as it has
a windows logo on most x86 tablets).

Fix the hwdb entry to only swap the buttons and not change the keycode
emitted by the home key.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>