service: remove pid watchers at sigchild when cgroup is gone
If a forking services cgroup quits before the watcher for the main pid is up,
the service becomes stale and the watched pid is blocked from getting recycled.
This situation is eventually resolved when the services timer expires, but
waiting for timer expiration delays recycling of the watched pid.
This can lead to pid starvation on large systems with lots of forking processes.

This patch tests at a sigchild event whether the cgroup still exists, and if
not, removes the watchers, so that the pid can quickly get re-used by other
services.