rules: do not ignore blank encrypted devices
Since the rule is not limited to add actions, it would bring down a mapper device (in systemd sense) if we perform wipefs (or even mkfs) on it. And since systemd-cryptsetup service binds to its mapper device (for cases like `cryptsetup close` I suppose), the service would be stopped and the mapper (in actual block device sense) would hence be gone.

It can be fixed by limiting the rule to add actions. However, fixing it that way would bring inconsistency. Let say we perform wipefs on a mapper (that has filesystem signature on it), it would remain "SYSTEMD_READY" before the systemd-cryptsetup service is restarted, but would become "not ready" after that.

Therefore, removing the rule.