fstab-generator: add Before= dependency to mount units using automount
The current practice of omitting the Before= dependency for
select mount units is both problematic and unnecessary, at least
for the automount case.

It is problematic because there is then no ordering imposed
at shutdown, so the mount units can be - and often are - stopped
before various services which might be using the mounted
filesystem.  This causes the unmount to fail, but allows shutdown
to proceed.
If the filesystem was an NFS filesystem, then once the network is
disabled it will become impossible to perform a final fsync,
and shutdown hangs indefinitely.

It is unnecessary in the automount case because there is no harm in
having the Before= dependency.
The mount unit will not even be started in the same transaction as
the *-fs.target, so - much as was the case for "noauto" mounts which
were also affected until recently (commit bd6bcce4d7fd9) - there is no
meaning to any ordering between the .mount and the *-fs.target.

Fixes #5991.