core/mount: ensure filesystems are unmounted at the correct time
If a mount is allowed to fail (e.g. marked "nofail" in /etc/fstab)
then it is likely that Before=local-fs.target would not be wanted,
as that could cause an unhelpful delay (and fstab-generator excludes
local-fs.target for that reason).

However, once the filesystem is mounted, it is important that it doesn't
get unmounted at shutdown until any services using it have completed,
otherwise the unmount might fail.

So at mount time, we add a Before dependency on local-fs.target.

Note that, since
 Commit: a6d305f91d72 ("mount: add remote-fs dependencies if needed
 after change")
we already auto-add a Before dependency on remote-fs.target for remote
filesystems.  This change just makes local mounts behave in the same
way.

Fixes #5991